+++ worker.js
@@ -1,4 +1,3 @@
-const { promisify } = require('util')
 const Queue = require('bull')
 const errors = require('request-promise-native/errors')
 
@@ -10,7 +9,6 @@ const { currentSource } = require('./src/sources')
 const logger = require('./src/logger')
 
 const REDIS_URL = (process.env.REDIS_URL !== undefined) ? process.env.REDIS_URL : 'redis://localhost:6379/1'
-const client = createClient()
 
 function handleError (job, err) {
   logger.error(err)
@@ -23,8 +21,9 @@ tokenQueue.on('failed', handleError)
 tokenQueue.process(requestToken)
 
 tokenQueue.on('completed', async function (job, result) {
-  const setAsync = promisify(client.set).bind(client)
-  await setAsync('access_token', result)
+  const client = createClient()
+  await client.set('access_token', result)
+  await client.quit()
   logger.info(`Successfully updated access token`)
 })
 
@@ -38,20 +37,21 @@ eventsQueue.on('failed', handleError)
 eventsQueue.process(processEvent)
 
 eventsQueue.on('active', async (job, jobPromise) => {
-  const sismemberAsync = promisify(client.sismember).bind(client)
-  let event = job.data
+  const client = createClient()
+  const event = job.data
 
-  const count = await sismemberAsync('processed_events', event.url)
+  const count = await client.sismember('processed_events', event.url)
   if (count) {
     logger.debug(`A event with url ${event.url} and name ${event.name} exists already.`)
     jobPromise.cancel()
   }
+  await client.quit()
 })
 
 eventsQueue.on('completed', async (job, result) => {
   let event = result
-  const saddAsync = promisify(client.sadd).bind(client)
-  await saddAsync('processed_events', event.url)
+  const client = createClient()
+  await client.sadd('processed_events', event.url)
   try {
     await createEvents([event], client)
     logger.info(`Added event with url ${event.url} and name ${event.name}`)
@@ -61,6 +61,8 @@ eventsQueue.on('completed', async (job, result) => {
     } else {
       logger.error(err)
     }
+  } finally {
+    await client.quit()
   }
 })
 
@@ -85,7 +87,10 @@ const currentSourceQueue = new Queue('current_source', REDIS_URL)
 currentSourceQueue.on('failed', handleError)
 
 currentSourceQueue.process(async () => {
-  return currentSource(client)
+  const client = createClient()
+  const source = await currentSource(client)
+  await client.quit()
+  return source
 })
 
 currentSourceQueue.on('completed', (job, result) => {
@@ -96,8 +101,3 @@ currentSourceQueue.on('completed', (job, result) => {
 currentSourceQueue.add({
   repeat: { cron: '*/10 0,7-21 * * *' }
 }) // Every 10 minutes.
-
-process.on('SIGTERM', () => {
-  client.quit()
-  logger.info(`Cleaning up redis connection.`)
-})